refactor: (#531) verified student api 이동 및 url 수정
