+++ dms-presentation/src/main/kotlin/team/aliens/dms/domain/file/FileWebAdapter.kt
@@ -1,20 +1,17 @@
 package team.aliens.dms.domain.file
 
-import org.springframework.http.HttpStatus
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RequestPart
-import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 import org.springframework.web.multipart.MultipartFile
 import team.aliens.dms.common.extension.toFile
 import team.aliens.dms.domain.file.dto.GetFileUploadUrlResponse
 import team.aliens.dms.domain.file.dto.response.UploadFileResponse
 import team.aliens.dms.domain.file.usecase.GetFileUploadUrlUseCase
-import team.aliens.dms.domain.file.usecase.ImportStudentUseCase
 import team.aliens.dms.domain.file.usecase.UploadFileUseCase
 import javax.validation.constraints.NotNull
 
@@ -23,8 +20,7 @@ import javax.validation.constraints.NotNull
 @RestController
 class FileWebAdapter(
     private val uploadFileUseCase: UploadFileUseCase,
-    private val getFileUploadUrlUseCase: GetFileUploadUrlUseCase,
-    private val importStudentUseCase: ImportStudentUseCase
+    private val getFileUploadUrlUseCase: GetFileUploadUrlUseCase
 ) {
 
     @PostMapping
@@ -40,13 +36,5 @@ class FileWebAdapter(
         @RequestParam("file_name") @NotNull fileName: String?
     ): GetFileUploadUrlResponse {
         return getFileUploadUrlUseCase.execute(fileName!!)
-    }
-
-    @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping("/verified-student")
-    fun importVerifiedStudentFromExcel(@RequestPart @NotNull file: MultipartFile?) {
-        importStudentUseCase.execute(
-            file!!.toFile()
-        )
     }
 }
